modified console menu

+++ HR.ConsoleApp/Program.cs
@@ -11,21 +11,23 @@ EmployeeService employeeService = new();
 bool runApp = true;
 while (runApp)
 {
+    Start:
     Console.WriteLine("Choose the option:");
     Console.WriteLine("1 - Create company\n" +
                       "2 - Show all departments of the company\n" +
-                      "3 - Delete company from system\n" +
+                      "3 - Delete company\n" +
                       "4 - Show all companies\n" +
                       "-----------------------------------------\n" +
                       "5 - Create department\n" +
                       "6 - Add employee to the department\n" +
                       "7 - Show department's employees\n" +
                       "8 - Configure department's name and employee limit\n" +
-                      "9 - Show all departments\n" +
+                      "9 - Delete department\n" +
+                      "10 - Show all departments\n" +
                       "-----------------------------------------\n" +
-                      "10 - Create employee\n" +
-                      "11 - Fire employee\n" +
-                      "12 - Show all employees in the system\n" +
+                      "11 - Create employee\n" +
+                      "12 - Fire employee\n" +
+                      "13 - Show all employees in the system\n" +
                       "-----------------------------------------\n" +
                       "0 - Exit");
     string? option = Console.ReadLine();
@@ -33,7 +35,7 @@ while (runApp)
     bool isInt = int.TryParse(option, out optionNumber);
     if (isInt)
     {
-        if (optionNumber >= 0 && optionNumber <= 12)
+        if (optionNumber >= 0 && optionNumber <= 13)
         {
             switch (optionNumber) 
             {
@@ -57,9 +59,31 @@ while (runApp)
                 case (int)MenuEnum.ShowAllDepartmentsInCo:
                     try
                     {
-                        Console.WriteLine("Enter company's name:");
-                        string? companyName = Console.ReadLine();
-                        companyService.GetAllDepartments(companyName);
+                        var checkShowDepInCo = companyService.CheckExistence();
+                        if (checkShowDepInCo is true) 
+                        {
+                            var checkDepEx = departmentService.CheckExistence();
+                            if (checkDepEx is true)
+                            {
+                                Console.WriteLine("Enter company's name:");
+                                string? companyName = Console.ReadLine();
+                                companyService.GetAllDepartments(companyName);
+                            }
+                            else
+                            {
+                                Console.ForegroundColor = ConsoleColor.DarkRed;
+                                Console.WriteLine("No department is found in the system, create department and link to the company ID.");
+                                Console.ResetColor();
+                                goto case (int)MenuEnum.CreateDepartment;
+                            }
+                        }
+                        else
+                        {
+                            Console.ForegroundColor = ConsoleColor.DarkRed;
+                            Console.WriteLine("No company is found in the system, first create a company.");
+                            Console.ResetColor();
+                            goto case (int)MenuEnum.CreateCompany;
+                        }
                     }
                     catch (Exception ex)
                     {
@@ -72,27 +96,49 @@ while (runApp)
                 case (int)MenuEnum.DeleteCompany:
                     try
                     {
-                        Console.WriteLine("Enter company's name:");
-                        string? companyName = Console.ReadLine();
-                        companyService.Delete(companyName);
+                        var checkDelCo = companyService.CheckExistence();
+                        if (checkDelCo is true) 
+                        {
+                            Console.WriteLine("Enter company's name:");
+                            string? companyName = Console.ReadLine();
+                            companyService.Delete(companyName);
+                        }
+                        else
+                        {
+                            Console.ForegroundColor = ConsoleColor.DarkRed;
+                            Console.WriteLine("No company is found in the system to delete. Firstly, create company.");
+                            Console.ResetColor();
+                            goto case (int)MenuEnum.CreateCompany;
+                        }       
                     }
                     catch (Exception ex)
                     {
                         Console.ForegroundColor = ConsoleColor.DarkRed;
                         Console.WriteLine(ex.Message);
                         Console.ResetColor();
-                        goto case (int)MenuEnum.DeleteCompany;
+                        goto Start;
                     }
                     break;
                 case (int)MenuEnum.ShowAllCompanies:
+                        var checkShowAllCo = companyService.CheckExistence();
+                    if (checkShowAllCo is true)
+                    {
                         Console.WriteLine("All companies listed below:");
                         companyService.ShowAll();
+                    }
+                    else
+                    {
+                        Console.ForegroundColor = ConsoleColor.DarkRed;
+                        Console.WriteLine("No company is found in the system.");
+                        Console.ResetColor();
+                        goto Start;
+                    }
                     break;
                 case (int)MenuEnum.CreateDepartment:
                     try
                     {
-                        var check = companyService.CheckExistence();
-                        if (check is true) 
+                        var checkCreateDep = companyService.CheckExistence();
+                        if (checkCreateDep is true) 
                         {
                             Console.WriteLine("Enter department name:");
                             string? departmentName = Console.ReadLine();
@@ -123,14 +169,25 @@ while (runApp)
                 case (int)MenuEnum.AddEmployee:
                     try
                     {
-                        var check = employeeService.CheckExistence();
-                        if (check is true)
+                        var checkAddEmployee = employeeService.CheckExistence();
+                        if (checkAddEmployee is true)
                         {
-                            Console.WriteLine("Enter employee's ID:");
-                            int employeeID = Convert.ToInt32(Console.ReadLine());
-                            Console.WriteLine("Enter department's ID:");
-                            int departmentID = Convert.ToInt32(Console.ReadLine());
-                            departmentService.AddEmployee(employeeID, departmentID);
+                            var checkDepExist = departmentService.CheckExistence();
+                            if (checkDepExist is true) 
+                            {
+                                Console.WriteLine("Enter employee's ID:");
+                                int employeeID = Convert.ToInt32(Console.ReadLine());
+                                Console.WriteLine("Enter department's ID:");
+                                int departmentID = Convert.ToInt32(Console.ReadLine());
+                                departmentService.AddEmployee(employeeID, departmentID);
+                            }
+                            else
+                            {
+                                Console.ForegroundColor = ConsoleColor.DarkRed;
+                                Console.WriteLine("No department is found in the system. Create department then add employees to the department.");
+                                Console.ResetColor();
+                                goto case (int)MenuEnum.CreateDepartment;
+                            }
                         }
                         else
                         {
@@ -180,6 +237,34 @@ while (runApp)
                         Console.WriteLine(ex.Message);
                         Console.ResetColor();
                         goto case (int)MenuEnum.UpdateDepartment;
+                    }
+                    break;
+                case (int)MenuEnum.DeleteDepartment:
+                    try
+                    {
+                        var checkDelDep = departmentService.CheckExistence();
+                        if (checkDelDep is true) 
+                        {
+                            Console.WriteLine("Enter department ID you want to delete:");
+                            int departmentId = Convert.ToInt32(Console.ReadLine());
+                            departmentService.DeleteDepartment(departmentId);
+                        }
+                        else
+                        {
+                            Console.ForegroundColor = ConsoleColor.DarkRed;
+                            Console.WriteLine($"Currently, no department is found in the system.\n" +
+                                              $"Firstly create company, if there is none in the system or create department directly to add this company.");
+                            Console.ResetColor();
+                            goto Start;
+                        }
+                        
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.ForegroundColor = ConsoleColor.DarkRed;
+                        Console.WriteLine(ex.Message);
+                        Console.ResetColor();
+                        goto case (int)MenuEnum.DeleteDepartment;
                     }
                     break;
                 case (int)MenuEnum.ShowAllDeps:
